Đã viết xong các chức năng của nhân viên và khách hàng. Còn phần tìm kiếm

+++ Views/QuanLy.cs
@@ -21,29 +21,60 @@ namespace QLThuCung.Views
         private void QuanLy_Load(object sender, EventArgs e)
         {
             panelLuuHuy.Hide();
+            panelLuuHuyKH.Hide();
+            panelLuuHuyNV.Hide();
         }
 
         void ClearTextBox()
         {
-            foreach (Control c in pnlDataBind.Controls)
+            if (pnlDataBind.Visible)
             {
-                if (c is TextBox)
+                foreach (Control c in pnlDataBind.Controls)
                 {
-                    TextBox questionTextBox = c as TextBox;
-                    if (questionTextBox != null)
+                    if (c is TextBox)
                     {
-                        questionTextBox.Text = "";
+                        TextBox questionTextBox = c as TextBox;
+                        if (questionTextBox != null)
+                        {
+                            questionTextBox.Text = "";
+                        }
+                    }
+                }
+            }
+            else if (pnlDataBindKH.Visible)
+            {
+                foreach (Control c in pnlDataBindKH.Controls)
+                {
+                    if (c is TextBox)
+                    {
+                        TextBox questionTextBox = c as TextBox;
+                        if (questionTextBox != null)
+                        {
+                            questionTextBox.Text = "";
+                        }
+                    }
+                }
+            }
+            else if (pnlDataBindNV.Visible)
+            {
+                foreach (Control c in pnlDataBindNV.Controls)
+                {
+                    if (c is TextBox)
+                    {
+                        TextBox questionTextBox = c as TextBox;
+                        if (questionTextBox != null)
+                        {
+                            questionTextBox.Text = "";
+                        }
                     }
                 }
             }
 
         }
-       
-
 
         private void tabctrlMain_Click(object sender, EventArgs e)
         {
-           if (tabctrlMain.SelectedTab == tabThuCung)
+            if (tabctrlMain.SelectedTab == tabThuCung)
             {
                 LoadThuCung();
                 tabctrlMain.TabPages.Remove(tabAdmin);
@@ -51,16 +82,22 @@ namespace QLThuCung.Views
             }
             else if (tabctrlMain.SelectedTab == tabNhanVien)
             {
+                LoadNhanVien();
+                tabctrlMain.TabPages.Remove(tabAdmin);
+                tabctrlMain.TabPages.Add(tabAdmin);
                 //db.usp_MuaDichVu("C101", "E001", "C001", DateTime.Today, 190000, "", "", "S003", "");
             }
             else if (tabctrlMain.SelectedTab == tabKhachHang)
             {
+                LoadKhachHang();
+                tabctrlMain.TabPages.Remove(tabAdmin);
+                tabctrlMain.TabPages.Add(tabAdmin);
                 //var result = from c in db.Customers select new { IDCus = c.ID_Cus, Name = c.Name, Phone = c.Phone };
-               // dgvThuCung.DataSource = result.ToList();
+                // dgvThuCung.DataSource = result.ToList();
             }
         }
 
-
+        #region TAB THU CUNG
         //======================Start THU CUNG============================
         void DataBindThuCung()
         {
@@ -69,7 +106,7 @@ namespace QLThuCung.Views
             tbLoaiThuCung.Text = dgvThuCung.CurrentRow.Cells[1].Value.ToString();
             //cbLoaiThuCung.SelectedIndex = cbLoaiThuCung.FindStringExact("");
             cbbTimThuCung.SelectedItem = "Loại";
-            cbGioitinh.SelectedItem = dgvThuCung.CurrentRow.Cells[2].Value.ToString(); 
+            cbGioitinh.SelectedItem = dgvThuCung.CurrentRow.Cells[2].Value.ToString();
             tbGiaNhapThuCung.Text = dgvThuCung.CurrentRow.Cells[3].Value.ToString();
             tbNccThuCung.Text = dgvThuCung.CurrentRow.Cells[4].Value.ToString();
             tbCanNangThuCung.Text = dgvThuCung.CurrentRow.Cells[5].Value.ToString();
@@ -108,7 +145,7 @@ namespace QLThuCung.Views
             try
             {
                 Pet pet = new Pet();
-   
+
                 pet.ID_Pet = tbIDThuCung.Text.ToString().Trim();
                 pet.ID_Spec = tbLoaiThuCung.Text.ToString().Trim();
                 pet.Sex = cbGioitinh.Text.ToString().Trim();
@@ -172,11 +209,14 @@ namespace QLThuCung.Views
         {
             try
             {
-                Pet pet = db.Pets.Where(p => p.ID_Pet == tbIDThuCung.Text.ToString()).SingleOrDefault();
-                db.Pets.Remove(pet);
-                db.SaveChanges();
-                MessageBox.Show("Delete Successfull!", "Thong Bao", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                LoadThuCung();
+                if (MessageBox.Show("Do you want delete this pet?", "Delete Pet", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                {
+                    Pet pet = db.Pets.Where(p => p.ID_Pet == tbIDThuCung.Text.ToString()).SingleOrDefault();
+                    db.Pets.Remove(pet);
+                    db.SaveChanges();
+                    MessageBox.Show("Delete Successfull!", "Thong Bao", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    LoadThuCung();
+                }
             }
             catch (Exception)
             {
@@ -203,5 +243,223 @@ namespace QLThuCung.Views
             dgvThuCung.DataSource = ms;
         }
         //======================End THU CUNG============================
+        #endregion
+
+        #region TAB KHACH HANG
+        void DataBindKhachHang()
+        {
+            txtMaKhachHang.Text = dgvKhachHang.CurrentRow.Cells[0].Value.ToString();
+            txtTenKhachHang.Text = dgvKhachHang.CurrentRow.Cells[1].Value.ToString();
+            cbbTimKhachHang.SelectedItem = "Tùy Chọn";
+            txtSDTKhachHang.Text = dgvKhachHang.CurrentRow.Cells[2].Value.ToString();
+            txtTuoiKhachHang.Text = dgvKhachHang.CurrentRow.Cells[3].Value.ToString();
+        }
+        void LoadKhachHang()
+        {
+            var result = from c in db.Customers select new { IDCus = c.ID_Cus, TenKhachHang = c.Name, SDT = c.Phone, Tuoi = c.Age };
+            dgvKhachHang.DataSource = result.ToList();
+            DataBindKhachHang();
+        }
+        private void dgvKhachHang_CellClick(object sender, DataGridViewCellEventArgs e)
+        {
+            DataBindKhachHang();
+        }
+        private void btnThemKhachHang_Click(object sender, EventArgs e)
+        {
+            ClearTextBox();
+            btnSuaKhachHang.Enabled = false;
+            btnXoaKhachHang.Enabled = false;
+            panelLuuHuyKH.Show();
+        }
+        private void btnSuaKhachHang_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                string idkh = dgvKhachHang.SelectedCells[0].OwningRow.Cells[0].Value.ToString().Trim();
+                Customer cus = db.Customers.Find(idkh);
+                cus.Name = txtTenKhachHang.Text.ToString().Trim();
+                cus.Phone = txtSDTKhachHang.Text.ToString().Trim();
+                cus.Age = Convert.ToInt32(txtTuoiKhachHang.Text);
+
+                db.SaveChanges();
+                MessageBox.Show("Update Successfull!", "Thong Bao", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                LoadKhachHang();
+            }
+            catch (Exception)
+            {
+                MessageBox.Show("Lỗi Không Update Được!", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                db = new ThuCungEntities();
+                LoadKhachHang();
+            }
+        }
+        private void btnXoaKhachHang_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                if (MessageBox.Show("Do you want delete this customer?", "Delete Customer", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                {
+                    Customer cus = db.Customers.Where(p => p.ID_Cus == txtMaKhachHang.Text.ToString()).SingleOrDefault();
+                    db.Customers.Remove(cus);
+                    db.SaveChanges();
+                    MessageBox.Show("Delete Successfull!", "Thong Bao", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    LoadKhachHang();
+                }
+            }
+            catch (Exception)
+            {
+                MessageBox.Show("Lỗi Không Delete Được!", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                db = new ThuCungEntities();
+                LoadKhachHang();
+            }
+        }
+        private void btnLuuKhachHang_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                Customer cus = new Customer();
+
+                cus.ID_Cus = txtMaKhachHang.Text.ToString().Trim();
+                cus.Name = txtTenKhachHang.Text.ToString().Trim();
+                cus.Phone = txtSDTKhachHang.Text.ToString().Trim();
+                cus.Age = Convert.ToInt32(txtTuoiKhachHang.Text);
+
+                db.Customers.Add(cus);
+                db.SaveChanges();
+
+                MessageBox.Show("Thêm Thành Công!", "Thong Bao", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                LoadKhachHang();
+                panelLuuHuyKH.Hide();
+                btnSuaKhachHang.Enabled = true;
+                btnXoaKhachHang.Enabled = true;
+            }
+            catch (Exception)
+            {
+                MessageBox.Show("Lỗi Không Thêm Được!", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                db = new ThuCungEntities();
+                LoadKhachHang();
+            }
+        }
+        private void btnHuyKhachHang_Click(object sender, EventArgs e)
+        {
+            LoadKhachHang();
+            panelLuuHuyKH.Hide();
+            btnSuaKhachHang.Enabled = true;
+            btnXoaKhachHang.Enabled = true;
+        }
+        #endregion
+
+        #region TAB NHAN VIEN
+        void DataBindNhanVien()
+        {
+            txtMaNhanVien.Text = dgvNhanVien.CurrentRow.Cells[0].Value.ToString();
+            txtTenNhanVien.Text = dgvNhanVien.CurrentRow.Cells[1].Value.ToString();
+            cbbTimNhanVien.SelectedItem = "Tùy Chọn";
+            txtSDTNhanVien.Text = dgvNhanVien.CurrentRow.Cells[2].Value.ToString();
+            txtTuoiNhanVien.Text = dgvNhanVien.CurrentRow.Cells[3].Value.ToString();
+            txtDiaChiNhanVien.Text = dgvNhanVien.CurrentRow.Cells[4].Value.ToString();
+            txtLuongNhanVien.Text = dgvNhanVien.CurrentRow.Cells[5].Value.ToString();
+            dtpNgayVaoLam.Text = dgvNhanVien.CurrentRow.Cells[6].Value.ToString();
+        }
+
+        void LoadNhanVien()
+        {
+            var result = from c in db.Employees select new { IDEmp = c.ID_Emp, TenNhanVien = c.Name, SDT = c.Phone, Tuoi = c.Age, DiaChi = c.Address, Luong = c.Salary, NgayVaoLam = c.DateStart };
+            dgvNhanVien.DataSource = result.ToList();
+            DataBindNhanVien();
+        }
+
+        private void dgvNhanVien_CellClick(object sender, DataGridViewCellEventArgs e)
+        {
+            DataBindNhanVien();
+        }
+        private void btnThemNhanVien_Click(object sender, EventArgs e)
+        {
+            ClearTextBox();
+            btnSuaNhanVien.Enabled = false;
+            btnXoaNhanVien.Enabled = false;
+            panelLuuHuyNV.Show();
+        }
+        private void btnSuaNhanVien_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                string idemp = dgvNhanVien.SelectedCells[0].OwningRow.Cells[0].Value.ToString().Trim();
+                Employee emp = db.Employees.Find(idemp);
+                emp.Name = txtTenNhanVien.Text.ToString().Trim();
+                emp.Phone = txtSDTNhanVien.Text.ToString().Trim();
+                emp.Age = Convert.ToInt32(txtTuoiNhanVien.Text);
+                emp.Address = txtDiaChiNhanVien.Text.ToString().Trim();
+                emp.Salary = Convert.ToInt32(txtLuongNhanVien.Text);
+                emp.DateStart = DateTime.Parse(dtpNgayVaoLam.Text);
+
+                db.SaveChanges();
+                MessageBox.Show("Update Successfull!", "Thong Bao", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                LoadNhanVien();
+            }
+            catch (Exception)
+            {
+                MessageBox.Show("Lỗi Không Update Được!", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                db = new ThuCungEntities();
+                LoadNhanVien();
+            }
+        }
+        private void btnXoaNhanVien_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                if (MessageBox.Show("Do you want delete this employee?", "Delete Employee", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                {
+                    Employee emp = db.Employees.Where(p => p.ID_Emp == txtMaNhanVien.Text.ToString()).SingleOrDefault();
+                    db.Employees.Remove(emp);
+                    db.SaveChanges();
+                    MessageBox.Show("Delete Successfull!", "Thong Bao", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    LoadNhanVien();
+                }
+            }
+            catch (Exception)
+            {
+                MessageBox.Show("Lỗi Không Delete Được!", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                db = new ThuCungEntities();
+                LoadNhanVien();
+            }
+        }
+        private void btnLuuNhanVien_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                Employee emp = new Employee();
+
+                emp.ID_Emp = txtMaNhanVien.Text.ToString().Trim();
+                emp.Name = txtTenNhanVien.Text.ToString().Trim();
+                emp.Phone = txtSDTNhanVien.Text.ToString().Trim();
+                emp.Age = Convert.ToInt32(txtTuoiNhanVien.Text);
+                emp.Address = txtDiaChiNhanVien.Text.ToString().Trim();
+                emp.Salary = Convert.ToInt32(txtLuongNhanVien.Text);
+                emp.DateStart = DateTime.Parse(dtpNgayVaoLam.Text);
+
+                db.Employees.Add(emp);
+                db.SaveChanges();
+
+                MessageBox.Show("Thêm Thành Công!", "Thong Bao", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                LoadNhanVien();
+                panelLuuHuyNV.Hide();
+                btnSuaNhanVien.Enabled = true;
+                btnXoaNhanVien.Enabled = true;
+            }
+            catch (Exception)
+            {
+                MessageBox.Show("Lỗi Không Thêm Được!", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                db = new ThuCungEntities();
+                LoadNhanVien();
+            }
+        }
+        private void btnHuyNhanVien_Click(object sender, EventArgs e)
+        {
+            LoadNhanVien();
+            panelLuuHuyNV.Hide();
+            btnSuaNhanVien.Enabled = true;
+            btnXoaNhanVien.Enabled = true;
+        }
+        #endregion
     }
 }
